requirements.txt
# Core LangGraph & AI Dependencies
langgraph>=0.0.32
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-core>=0.1.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Data & Storage
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
json5
typing-extensions
pyyaml>=6.0

# Monitoring & Logging
structlog>=23.0.0